added static kotlinter tool, some fixes in xml files, rename of main project package
