+++ app/build.gradle
@@ -1,10 +1,8 @@
 apply plugin: 'com.android.application'
-
 apply plugin: 'kotlin-android'
-
 apply plugin: 'kotlin-android-extensions'
-
 apply plugin: 'kotlin-kapt'
+apply plugin: 'org.jmailen.kotlinter'
 
 android {
     compileSdkVersion 28
@@ -27,7 +25,7 @@ android {
     }
 }
 ext {
-    supportLibraryVersion = '28.0.0-beta01'
+    supportLibraryVersion = '28.0.0-rc01'
     retrofitVersion = '2.4.0'
     kotlin_version = '1.2.60'
     junitVersion = '4.12'
@@ -44,18 +42,13 @@ dependencies {
     testImplementation "junit:junit:$junitVersion"
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    kapt 'com.android.databinding:compiler:3.1.3'
     implementation "com.squareup.retrofit2:retrofit:$retrofitVersion"
     implementation "com.squareup.retrofit2:converter-gson:$retrofitVersion"
     implementation "com.squareup.retrofit2:adapter-rxjava2:$retrofitVersion"
     implementation "com.android.support:recyclerview-v7:$supportLibraryVersion"
-    implementation'io.reactivex.rxjava2:rxjava:2.1.1'
-    implementation'io.reactivex.rxjava2:rxandroid:2.0.1'
+    implementation 'io.reactivex.rxjava2:rxjava:2.1.9'
+    implementation 'io.reactivex.rxjava2:rxandroid:2.0.2'
     implementation "com.squareup.picasso:picasso:$picassoVersion"
     implementation "com.android.support:cardview-v7:$supportLibraryVersion"
     implementation "android.arch.lifecycle:extensions:$lifecycleVersion"
-    implementation "android.arch.lifecycle:viewmodel:$lifecycleVersion"
-}
-kapt {
-    generateStubs = true
 }
