--- app/src/main/java/com/example/marcinwisniewski/intive_training_app/model/Movie.kt
@@ -1,9 +1,0 @@
-package com.example.marcinwisniewski.intive_training_app.model
-
-class Movie(id: Int, title: String, year: String, poster: String) {
-
-    var id: Int = id
-    var title: String? = title
-    var year: String? = year
-    var poster: String? = poster
-}
