--- app/src/main/java/com/example/marcinwisniewski/intive_training_app/model/MovieResponse.kt
@@ -1,6 +1,0 @@
-package com.example.marcinwisniewski.intive_training_app.model
-
-class MovieResponse{
-    var results :List<Movie>? = null
-
-}
