--- app/src/main/java/com/example/marcinwisniewski/intive_training_app/viewmodel/DetailsViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.marcinwisniewski.intive_training_app.viewmodel
-
-import android.arch.lifecycle.ViewModel
-
-
-class DetailsViewModel : ViewModel() {
-}
