--- app/src/main/java/com/example/marcinwisniewski/intive_training_app/viewmodel/ListViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.marcinwisniewski.intive_training_app.viewmodel
-
-import android.arch.lifecycle.ViewModel
-
-
-class ListViewModel() : ViewModel() {
-
-    //fun getMovies(): Observable<MovieResponse> {
-      //  this function will return the movie list
-    //}
-
-
-}
