--- app/src/main/java/com/example/marcinwisniewski/intive_training_app/viewmodel/SingleItemViewModel.kt
@@ -1,8 +1,0 @@
-package com.example.marcinwisniewski.intive_training_app.viewmodel
-
-import android.arch.lifecycle.ViewModel
-
-class SingleItemViewModel() : ViewModel() {
-
-
-}
