+++ app/src/main/java/com/example/marcinwisniewski/intivetrainingapp/model/Movie.kt
@@ -1,0 +1,7 @@
+package com.example.marcinwisniewski.intivetrainingapp.model
+
+class Movie(var id: Int, title: String, year: String, poster: String) {
+    var title: String? = title
+    var year: String? = year
+    var poster: String? = poster
+}
