+++ app/src/main/java/com/example/marcinwisniewski/intivetrainingapp/model/MovieResponse.kt
@@ -1,0 +1,5 @@
+package com.example.marcinwisniewski.intivetrainingapp.model
+
+class MovieResponse {
+    var results: List<Movie>? = null
+}
