+++ app/src/main/java/com/example/marcinwisniewski/intivetrainingapp/view/DetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.marcinwisniewski.intive_training_app.view
+package com.example.marcinwisniewski.intivetrainingapp.view
 
 import android.arch.lifecycle.ViewModelProviders
 import android.databinding.DataBindingUtil
@@ -7,15 +7,12 @@ import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.marcinwisniewski.intive_training_app.R
-import com.example.marcinwisniewski.intive_training_app.databinding.DetailsFragmentBinding
-import com.example.marcinwisniewski.intive_training_app.viewmodel.DetailsViewModel
-
+import com.example.marcinwisniewski.intivetrainingapp.R
+import com.example.marcinwisniewski.intivetrainingapp.databinding.DetailsFragmentBinding
+import com.example.marcinwisniewski.intivetrainingapp.viewmodel.DetailsViewModel
 
 class DetailsFragment : Fragment() {
-
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-
         val detailsBinding: DetailsFragmentBinding = DataBindingUtil.inflate(layoutInflater, R.layout.details_fragment, container, false)
         detailsBinding.viewModel = ViewModelProviders.of(this).get(DetailsViewModel::class.java)
         return detailsBinding.root
