+++ app/src/main/java/com/example/marcinwisniewski/intivetrainingapp/view/ListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.marcinwisniewski.intive_training_app.view
+package com.example.marcinwisniewski.intivetrainingapp.view
 
 import android.arch.lifecycle.ViewModelProviders
 import android.databinding.DataBindingUtil
@@ -7,14 +7,12 @@ import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.marcinwisniewski.intive_training_app.R
-import com.example.marcinwisniewski.intive_training_app.databinding.ListFragmentBinding
-import com.example.marcinwisniewski.intive_training_app.viewmodel.ListViewModel
+import com.example.marcinwisniewski.intivetrainingapp.R
+import com.example.marcinwisniewski.intivetrainingapp.databinding.ListFragmentBinding
+import com.example.marcinwisniewski.intivetrainingapp.viewmodel.ListViewModel
 
 class ListFragment : Fragment() {
-
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-
         val listBinding: ListFragmentBinding = DataBindingUtil.inflate(layoutInflater, R.layout.list_fragment, container, false)
         listBinding.viewModel = ViewModelProviders.of(this).get(ListViewModel::class.java)
         return listBinding.root
