+++ app/src/main/java/com/example/marcinwisniewski/intivetrainingapp/viewmodel/DetailsViewModel.kt
@@ -1,0 +1,5 @@
+package com.example.marcinwisniewski.intivetrainingapp.viewmodel
+
+import android.arch.lifecycle.ViewModel
+
+class DetailsViewModel : ViewModel()
