+++ app/src/main/java/com/example/marcinwisniewski/intivetrainingapp/viewmodel/ListViewModel.kt
@@ -1,0 +1,5 @@
+package com.example.marcinwisniewski.intivetrainingapp.viewmodel
+
+import android.arch.lifecycle.ViewModel
+
+class ListViewModel : ViewModel()
