+++ app/src/main/java/com/example/marcinwisniewski/intivetrainingapp/viewmodel/SingleItemViewModel.kt
@@ -1,0 +1,5 @@
+package com.example.marcinwisniewski.intivetrainingapp.viewmodel
+
+import android.arch.lifecycle.ViewModel
+
+class SingleItemViewModel : ViewModel()
