+++ app/src/main/res/layout/list_fragment.xml
@@ -2,13 +2,10 @@
 <layout>
 
     <data>
-
         <import type="android.view.View" />
-
         <variable
             name="viewModel"
-            type="com.example.marcinwisniewski.intive_training_app.viewmodel.ListViewModel" />
-
+            type="com.example.marcinwisniewski.intivetrainingapp.viewmodel.ListViewModel" />
     </data>
 
     <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
@@ -19,8 +16,8 @@
             android:id="@+id/recycler_list"
             android:layout_width="match_parent"
             android:layout_height="match_parent">
-
         </android.support.v7.widget.RecyclerView>
+
     </android.support.constraint.ConstraintLayout>
 
 </layout>
