+++ app/src/main/res/layout/list_row.xml
@@ -1,18 +1,15 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android">
 
     <data>
-
         <import type="android.view.View" />
-
         <variable
             name="listViewModel"
-            type="com.example.marcinwisniewski.intive_training_app.viewmodel.SingleItemViewModel" />
+            type="com.example.marcinwisniewski.intivetrainingapp.viewmodel.SingleItemViewModel" />
     </data>
 
-
-    <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
-
     </android.support.constraint.ConstraintLayout>
+
 </layout>
