+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">intive_training_app</string>
+    <string name="app_name">intivetrainingapp</string>
 </resources>
